Fixed RSSI, LQI and DR measurements

+++ openstack/02b-MAChigh/neighbors.h
@@ -35,7 +35,7 @@ typedef struct {
    int8_t           rssi;
    uint16_t          numRx;
    uint16_t          numTx;
-   uint8_t          numTxACK;
+   uint16_t          numTxACK;
    uint8_t          numWraps;//number of times the tx counter wraps. can be removed if memory is a restriction. also check openvisualizer then.
    asn_t            asn;
    uint8_t          joinPrio;
